pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "trans_image"
version = "0.1.0"
description = "AI-based image text automatic translation application"
requires-python = ">=3.11"
dependencies = [
    # GUI
    "PySide6>=6.7.0",
    "qasync>=0.27.0",
    # OCR
    "easyocr>=1.7.2",
    # Image processing
    "opencv-python>=4.10.0",
    "Pillow>=10.4.0",
    "fonttools>=4.53.0",
    "numpy>=1.26.0",
    # Language detection
    "lingua-language-detector>=2.0.2",
    # Translation APIs
    "deepl>=1.18.0",
    "google-genai>=1.0.0",
    "openai>=1.40.0",
    "requests>=2.32.0",
    "ollama>=0.3.0",
    # Agent APIs
    "anthropic>=0.34.0",
    # Config / utils
    "PyYAML>=6.0.2",
    "pydantic>=2.8.0",
    "asyncio-throttle>=1.0.2",
]

[project.optional-dependencies]
paddleocr = [
    "paddlepaddle>=2.6.1",
    "paddleocr>=2.8.1",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-qt>=4.4.0",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
]

[project.scripts]
trans-image = "trans_image.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "trans_image*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
